LayoutTests/imported/w3c/web-platform-tests/css/css-font-loading/fontface-loading-events-font-from-arraybuffer.html

<!DOCTYPE html>
<title>Test font loading from an ArrayBuffer</title>
<link rel="help" href="https://drafts.csswg.org/css-font-loading/#font-face-constructor">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
.initial {
  font-family: "AhemTest", sans-serif;
  font-size: 20px;
}
</style>
<div>Font loading test</div>
<script>  
  let expectedDoneFace = "AhemTest";
  let expectedErrorFace = "";

  const fontSet = document.fonts;

  promise_test(t => {
    return fontSet.ready;
  }, 'The initial ready promise fired');

  promise_test(async t => {
    const eventPromise = new Promise(resolve => {
      fontSet.addEventListener('loadingdone', t.step_func(event => {
        assert_equals(event.fontfaces.length, 1, `'loadingdone' FontFaceSet should have one font; it has ${event.fontfaces.length}`);
        assert_equals(event.fontfaces[0].family, expectedDoneFace, `'loadingdone' FontFaceSet fontFamily is ${expectedDoneFace}`);
        resolve();
      }));
    }, { once: true });

    const response = await fetch('/fonts/Ahem.ttf');
    const buffer = await response.arrayBuffer();

    expectedDoneFace = 'data-font';
    const dataFont = new FontFace('data-font', buffer);
    
    assert_equals(dataFont.status, "unloaded", "Initial status should be 'unloaded'");
    fontSet.add(dataFont);
    dataFont.load();

    return Promise.all([fontSet.ready, eventPromise]);
  }, 'Loading a font from an ArrayBuffer should fire a `loadingdone` event with the correct `fontfaceset`');

</script>
